Remove false warning of missing libpthread.so on Linux (a cherry-pick of commits ffd2f9fe0, 6c47e6d4c from 'master')
Issue #477 (bdwgc).

Some libc implementations like bionic, musl and glibc 2.34 (and later)
do not have libpthread.so because the pthreads-related code is located
in libc.so, thus potential calloc calls from such code are forwarded
to real (libc) calloc without any special handling on the libgc side.

* malloc.c [REDIRECT_MALLOC && !REDIRECT_MALLOC_IN_HEADER
&& GC_LINUX_THREADS] (GC_init_lib_bounds): Do not call WARN to report
missing libpthread.so if not defined __GLIBC__ or glibc
version is 2.34 or above; add comment and TODO item.