Fix reliance on glibc-specific `__data_start` for some Linux platforms
Setting `DATASTART` based on `__data_start` only works when the
`__data_start` symbol is defined somewhere, and this seems to be a
glibc-specific feature (i.e. not available on musl). To avoid a
"Wrong DATASTART/END pair" runtime error when building for non-glibc
Linux platforms, we should fall back to `SEARCH_FOR_DATA_START` if not
using glibc.