Fix reliance on glibc-specific '__data_start' for some Linux platforms
Setting `DATASTART` based on `__data_start` only works when the
`__data_start` symbol is defined somewhere, and this seems to be
a glibc-specific feature (e.g., not available on musl).

To avoid a "Wrong DATASTART/END pair" runtime error when building
for non-glibc Linux platforms, we should fall back to
`SEARCH_FOR_DATA_START` way.

* include/private/gcconfig.h [(LOONGARCH || NIOS2 || OR1K || S390
|| AARCH64 || ARC || TILEPRO || TILEGX || RISCV) && LINUX
&& !__GLIBC__] (__data_start): Do not declare.
* include/private/gcconfig.h [(LOONGARCH || NIOS2 || OR1K || S390
|| AARCH64 || ARC || TILEPRO || TILEGX || RISCV) && LINUX
&& !__GLIBC__] (DATASTART): Do not define.
* include/private/gcconfig.h [(LOONGARCH || NIOS2 || OR1K || S390
|| AARCH64 || ARC || TILEPRO || TILEGX || RISCV) && LINUX
&& !__GLIBC__] (SEARCH_FOR_DATA_START): Define.
* include/private/gcconfig.h [AARCH64 && LINUX]
(SEARCH_FOR_DATA_START): Do not check `HOST_ANDROID` macro (because
`__GLIBC__` is not defined on Android).