fix: unprefixed cmake options
The cmake options defined in the root project build script have very
generic names which can easily clash with other cache variables when
bdwgc is included as a subdirectory in a larger project.

This patch prefixes the cmake options with bdwgc_ and introduces a
deprecation system to warn clients about the change without breaking
their builds.