Don't call GC_delete_thread on DLL_THREAD_DETACH
This can cause race conditions with GC_stop_world thread suspension. It
might interleave like this:

1. Main begins suspending
2. Child thread sets handle to NULL
3. Main finishes suspending

This leaves us in a broken state, where main has no way of resuming the
thread, and the child thread has no way of actually closing its handle.

With this patch, DllMain no longer calls GC_delete_thread, but instead
marks it as ready to delete. It will be deleted instead by the main
thread.

This allows a lot of the old stop-gap code to be removed.