Fix 'reserved' argument handling in DllMain (a cherry-pick of commit f389c8f20 from 'release-7_4')
PR #830 (bdwgc).

Do not close thread handles on `DllMain` process exit.

A non-null `reserved` indicates that the process is exiting, in which
case the process handles will be closed automatically, so we should not
do that here.

* win32_threads.c [!GC_PTHREADS && !GC_NO_THREADS_DISCOVERY] (DllMain):
If `reserved` is non-null, then do not call `GC_delete_thread()`,
`GC_deinit()` and `DeleteCriticalSection()`; add comment.