Fix 'reserved' argument handling in GC_DllMain
PR #830 (bdwgc).

Do not close thread handles on `DllMain` process exit.

A non-null `reserved` indicates that the process is exiting, in which
case the process handles will be closed automatically, so we should not
do that here.

* win32_threads.c [!GC_NO_THREADS_DISCOVERY] (GC_DllMain): Do not
mark `reserved` as unused; if `reserved` is non-null, then do not
call `GC_delete_thread()` and `GC_deinit()`; add comment.